+++ src/selection/DistanceFinder.cpp
@@ -142,7 +142,7 @@ namespace OpenMD {
     SelectionSet bsTemp = bs;
     bsTemp              = bsTemp.parallelReduce();
 
-    for (int i = 0; i < bsTemp.bitsets_[STUNTDOUBLE].size(); ++i) {
+    for (size_t i = 0; i < bsTemp.bitsets_[STUNTDOUBLE].size(); ++i) {
       if (bsTemp.bitsets_[STUNTDOUBLE][i]) {
 #ifdef IS_MPI
 
@@ -168,7 +168,7 @@ namespace OpenMD {
         center    = stuntdoubles_[i];
         centerPos = center->getPos();
 #endif
-        for (unsigned int j = 0; j < molecules_.size(); ++j) {
+        for (size_t j = 0; j < molecules_.size(); ++j) {
           if (molecules_[j] != NULL) {
             Vector3d r = centerPos - molecules_[j]->getCom();
             currSnapshot->wrapVector(r);
@@ -177,7 +177,7 @@ namespace OpenMD {
             }
           }
         }
-        for (unsigned int j = 0; j < stuntdoubles_.size(); ++j) {
+        for (size_t j = 0; j < stuntdoubles_.size(); ++j) {
           if (stuntdoubles_[j] != NULL) {
             Vector3d r = centerPos - stuntdoubles_[j]->getPos();
             currSnapshot->wrapVector(r);
@@ -186,7 +186,7 @@ namespace OpenMD {
             }
           }
         }
-        for (unsigned int j = 0; j < bonds_.size(); ++j) {
+        for (size_t j = 0; j < bonds_.size(); ++j) {
           if (bonds_[j] != NULL) {
             Vector3d loc = bonds_[j]->getAtomA()->getPos();
             loc += bonds_[j]->getAtomB()->getPos();
@@ -196,7 +196,7 @@ namespace OpenMD {
             if (r.length() <= distance) { bsResult.bitsets_[BOND].setBitOn(j); }
           }
         }
-        for (unsigned int j = 0; j < bends_.size(); ++j) {
+        for (size_t j = 0; j < bends_.size(); ++j) {
           if (bends_[j] != NULL) {
             Vector3d loc = bends_[j]->getAtomA()->getPos();
             loc += bends_[j]->getAtomB()->getPos();
@@ -207,7 +207,7 @@ namespace OpenMD {
             if (r.length() <= distance) { bsResult.bitsets_[BEND].setBitOn(j); }
           }
         }
-        for (unsigned int j = 0; j < torsions_.size(); ++j) {
+        for (size_t j = 0; j < torsions_.size(); ++j) {
           if (torsions_[j] != NULL) {
             Vector3d loc = torsions_[j]->getAtomA()->getPos();
             loc += torsions_[j]->getAtomB()->getPos();
@@ -221,7 +221,7 @@ namespace OpenMD {
             }
           }
         }
-        for (unsigned int j = 0; j < inversions_.size(); ++j) {
+        for (size_t j = 0; j < inversions_.size(); ++j) {
           if (inversions_[j] != NULL) {
             Vector3d loc = inversions_[j]->getAtomA()->getPos();
             loc += inversions_[j]->getAtomB()->getPos();
@@ -268,7 +268,7 @@ namespace OpenMD {
     SelectionSet bsTemp = bs;
     bsTemp              = bsTemp.parallelReduce();
 
-    for (int i = 0; i < bsTemp.bitsets_[STUNTDOUBLE].size(); ++i) {
+    for (size_t i = 0; i < bsTemp.bitsets_[STUNTDOUBLE].size(); ++i) {
       if (bsTemp.bitsets_[STUNTDOUBLE][i]) {
         // Now, if we own stuntdouble i, we can use the position, but in
         // parallel, we'll need to let everyone else know what that
@@ -293,7 +293,7 @@ namespace OpenMD {
         center    = stuntdoubles_[i];
         centerPos = center->getPos(frame);
 #endif
-        for (unsigned int j = 0; j < molecules_.size(); ++j) {
+        for (size_t j = 0; j < molecules_.size(); ++j) {
           if (molecules_[j] != NULL) {
             Vector3d r = centerPos - molecules_[j]->getCom(frame);
             currSnapshot->wrapVector(r);
@@ -303,7 +303,7 @@ namespace OpenMD {
           }
         }
 
-        for (unsigned int j = 0; j < stuntdoubles_.size(); ++j) {
+        for (size_t j = 0; j < stuntdoubles_.size(); ++j) {
           if (stuntdoubles_[j] != NULL) {
             Vector3d r = centerPos - stuntdoubles_[j]->getPos(frame);
             currSnapshot->wrapVector(r);
@@ -312,7 +312,7 @@ namespace OpenMD {
             }
           }
         }
-        for (unsigned int j = 0; j < bonds_.size(); ++j) {
+        for (size_t j = 0; j < bonds_.size(); ++j) {
           if (bonds_[j] != NULL) {
             Vector3d loc = bonds_[j]->getAtomA()->getPos(frame);
             loc += bonds_[j]->getAtomB()->getPos(frame);
@@ -322,7 +322,7 @@ namespace OpenMD {
             if (r.length() <= distance) { bsResult.bitsets_[BOND].setBitOn(j); }
           }
         }
-        for (unsigned int j = 0; j < bends_.size(); ++j) {
+        for (size_t j = 0; j < bends_.size(); ++j) {
           if (bends_[j] != NULL) {
             Vector3d loc = bends_[j]->getAtomA()->getPos(frame);
             loc += bends_[j]->getAtomB()->getPos(frame);
@@ -333,7 +333,7 @@ namespace OpenMD {
             if (r.length() <= distance) { bsResult.bitsets_[BEND].setBitOn(j); }
           }
         }
-        for (unsigned int j = 0; j < torsions_.size(); ++j) {
+        for (size_t j = 0; j < torsions_.size(); ++j) {
           if (torsions_[j] != NULL) {
             Vector3d loc = torsions_[j]->getAtomA()->getPos(frame);
             loc += torsions_[j]->getAtomB()->getPos(frame);
@@ -347,7 +347,7 @@ namespace OpenMD {
             }
           }
         }
-        for (unsigned int j = 0; j < inversions_.size(); ++j) {
+        for (size_t j = 0; j < inversions_.size(); ++j) {
           if (inversions_[j] != NULL) {
             Vector3d loc = inversions_[j]->getAtomA()->getPos(frame);
             loc += inversions_[j]->getAtomB()->getPos(frame);
